wcAbst

make word cloud or correlation network from PubMed

Usage

wcAbst(
  queries,
  redo = NULL,
  madeUpper = c("dna", "rna"),
  target = "abstract",
  usefil = NA,
  filnum = 0,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  delim = "OR",
  geneUpper = TRUE,
  apiKey = NULL,
  tfidf = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "wc",
  colorText = FALSE,
  quote = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = FALSE,
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  bn = FALSE,
  R = 20,
  retMax = 10,
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  genePlot = FALSE,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  orgDb = org.Hs.eg.db,
  preset = FALSE,
  onWholeDTM = FALSE,
  madeUpperGenes = TRUE,
  ...
)

Arguments

queries

gene symbols (max: 5)

redo

if plot in other parameters, input the previous list

madeUpper

make the words uppercase in resulting plot

target

"abstract" or "title"

usefil

filter based on "gstfidf" (whole gene summary tf-idf) or "bsdbtfidf" (whole bugsigdb tf-idf)

filnum

specify filter tfidf

pvclAlpha

alpha for pvpick()

numOnly

delete number only

delim

delimiter for queries

geneUpper

make queries uppercase

apiKey

api key for eutilities

tfidf

use TfIdf when making TDM

pal

palette for color gradient in correlation network

numWords

the number of words to be shown

scaleRange

scale for label and node size in correlation network

showLegend

whether to show legend in correlation network

plotType

"wc" or "network"

colorText

color text label based on frequency in correlation network

quote

whether to quote the queries

corThresh

the correlation threshold

layout

the layout for correlation network, defaul to "nicely"

tag

cluster the words based on text using pvclust

tagWhole

tag based on whole data or subset

onlyCorpus

return only corpus

onlyTDM

return only TDM

bn

perform bootstrap-based Bayesian network inference instead of correlation using bnlearn

R

how many bootstrap when bn is stated

retMax

how many items are to be retlieved?

edgeLabel

if TRUE, plot the edge label (default: FALSE)

edgeLink

if FALSE, use geom_edge_diagonal

ngram

default to NA (1)

genePlot

plot associated genes (default: FALSE)

deleteZeroDeg

delete zero degree node from plot in correlation network

additionalRemove

specific words to be excluded

orgDb

org database, default to org.Hs.eg.db

preset

filter preset words

onWholeDTM

calculate correlation network on whole dataset or top-words specified by numWords

madeUpperGenes

make genes upper case automatically (default to TRUE)

...

parameters to pass to wordcloud()

Value

list of data frame and ggplot2 object

Examples

wcAbst("DDX41")
#> proceeding without API key
#> proceeding without API key
#> type: pubmed_abstract
#> DDX41
#> 197.4 Kb