wcBSDB

Visualize BugSigDB

Usage

wcBSDB(
  mbList,
  excludeFreq = 1000,
  excludeTfIdf = NA,
  additionalRemove = NA,
  tfidf = FALSE,
  target = "title",
  apiKey = NULL,
  pre = FALSE,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  madeUpper = c("dna", "rna"),
  redo = NULL,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  edgeLabel = FALSE,
  mbPlot = FALSE,
  ngram = NA,
  plotType = "wc",
  disPlot = FALSE,
  onWholeDTM = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  tag = FALSE,
  tagWhole = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  cl = FALSE,
  ...
)

Arguments

mbList

microbe list

excludeFreq

exclude words with overall frequency above excludeFreq default to 0

excludeTfIdf

exclude based on tfidf (default: NA)

additionalRemove

specific words to be excluded

tfidf

use TfIdf when making TDM

target

"title" or "abstract"

apiKey

api key for eutilities

pre

predefined filter words

pvclAlpha

alpha for pvpick()

numOnly

delete number only

madeUpper

make the words uppercase in resulting plot

redo

if "abstract" is chosen in target, one can provide resulting object again

pal

palette for color gradient in correlation network and tag coloring

numWords

the number of words to be shown

scaleRange

scale for label and node size in correlation network

showLegend

whether to show legend in correlation network

edgeLabel

if TRUE, plot the edge label (default: FALSE)

mbPlot

plot microbe names

ngram

default to NA (1)

plotType

"wc" or "network"

disPlot

plot diseases

onWholeDTM

calculate correlation network on whole dataset or top-words specified by numWords

colorText

color text label based on frequency in correlation network

corThresh

the correlation threshold

tag

perform pvclust on words and colorlize them in wordcloud

tagWhole

tag whole set or subset

layout

the layout for correlation network, defaul to "nicely"

edgeLink

if FALSE, use geom_edge_diagonal

deleteZeroDeg

delete zero degree node from plot in correlation network

cl

cluster object passed to pvclust

...

parameters to pass to wordcloud()

Value

list of data frame and ggplot2 object

Examples

mbList <- c("Veillonella dispar")
wcBSDB(mbList)
#> Input microbes: 1
#> Error in completeSubclasses(classDef2, class1, obj, where) : 
#>   trying to get slot "subclasses" from an object of a basic class ("NULL") with no slots
#> Error: .onLoad failed in loadNamespace() for 'dbplyr', details:
#>   call: setClass(cl, contains = c(prevClass, "VIRTUAL"), where = where)
#>   error: error in contained classes ("character") for class "ident"; class definition removed from 'dbplyr'