getWordsOnDendro

Get grobs to plot on the dendrogram with the position information

Usage

getWordsOnDendro(
  dhc,
  geneVec,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  excludeTfIdf = NA,
  numberOfWords = 25,
  excludeFreq = 5000,
  orgDb = org.Hs.eg.db,
  tfidf = FALSE,
  madeUpper = c("rna", "dna")
)

Arguments

dhc

dendrogram

geneVec

gene-named vector of node names in dendrogram

geneNumLimit

when the gene number is above this threshold, pyramid plots are not produced

geneVecType

type of the name of geneVec (default: ENSEMBL)

excludeTfIdf

filter using tfidf

numberOfWords

the number of words to plot (default: 25)

excludeFreq

words with the frequency above this threshold is excluded beforehand

orgDb

organism database

tfidf

use tfidf when making TDM

madeUpper

words with uppercase

Value

list of pyramid plot grobs and its positions

Examples

mod <- returnExample()
#> 'select()' returned 1:many mapping between keys and columns
#> 'select()' returned 1:1 mapping between keys and columns
#> 'select()' returned 1:1 mapping between keys and columns
result <- hclust(dist(t(mod$MEs)))
dhc <- result |>
    as.dendrogram() |>
    dendextend::hang.dendrogram()
geneVec <- paste0("ME", mod$colors)
names(geneVec) <- names(mod$colors)
getWordsOnDendro(dhc, geneVec, numberOfWords = 2)
#> 'select()' returned 1:1 mapping between keys and columns
#> 'select()' returned 1:1 mapping between keys and columns
#> $`1`
#> $`1`$plot
#> TableGrob (20 x 47) "layout": 52 grobs
#>     z         cells       name                                       grob
#> 1   5 ( 9- 9, 7- 7)   spacer-1                             zeroGrob[NULL]
#> 2   7 (11-11, 7- 7)   axis-l-1         absoluteGrob[GRID.absoluteGrob.23]
#> 3   3 (13-13, 7- 7)   spacer-1                             zeroGrob[NULL]
#> 4   6 ( 9- 9, 9- 9)   axis-t-1                             zeroGrob[NULL]
#> 5   1 (11-11, 9- 9)    panel-1                    gTree[panel-1.gTree.21]
#> 6   9 (13-13, 9- 9)   axis-b-1         absoluteGrob[GRID.absoluteGrob.22]
#> 7   4 ( 9- 9,11-11)   spacer-1                             zeroGrob[NULL]
#> 8   8 (11-11,11-11)   axis-r-1                             zeroGrob[NULL]
#> 9   2 (13-13,11-11)   spacer-1                             zeroGrob[NULL]
#> 10 10 ( 8- 8, 9- 9)   xlab-t-1                             zeroGrob[NULL]
#> 11 11 (14-14, 9- 9)   xlab-b-1 zeroGrob[axis.title.x.bottom..zeroGrob.24]
#> 12 12 (11-11, 6- 6)   ylab-l-1   zeroGrob[axis.title.y.left..zeroGrob.25]
#> 13 13 (11-11,12-12)   ylab-r-1                             zeroGrob[NULL]
#> 14 14 ( 5- 5, 9- 9) subtitle-1       zeroGrob[plot.subtitle..zeroGrob.27]
#> 15 15 ( 4- 4, 9- 9)    title-1          zeroGrob[plot.title..zeroGrob.26]
#> 16 16 (15-15, 9- 9)  caption-1        zeroGrob[plot.caption..zeroGrob.29]
#> 17 17 ( 3- 3, 3- 3)      tag-1            zeroGrob[plot.tag..zeroGrob.28]
#> 18 22 ( 9- 9,22-22)   spacer-2                             zeroGrob[NULL]
#> 19 24 (11-11,22-22)   axis-l-2         absoluteGrob[GRID.absoluteGrob.41]
#> 20 20 (13-13,22-22)   spacer-2                             zeroGrob[NULL]
#> 21 23 ( 9- 9,24-24)   axis-t-2                             zeroGrob[NULL]
#> 22 18 (11-11,24-24)    panel-2                    gTree[panel-1.gTree.39]
#> 23 26 (13-13,24-24)   axis-b-2         absoluteGrob[GRID.absoluteGrob.40]
#> 24 21 ( 9- 9,26-26)   spacer-2                             zeroGrob[NULL]
#> 25 25 (11-11,26-26)   axis-r-2                             zeroGrob[NULL]
#> 26 19 (13-13,26-26)   spacer-2                             zeroGrob[NULL]
#> 27 27 ( 8- 8,24-24)   xlab-t-2                             zeroGrob[NULL]
#> 28 28 (14-14,24-24)   xlab-b-2 zeroGrob[axis.title.x.bottom..zeroGrob.42]
#> 29 29 (11-11,21-21)   ylab-l-2   zeroGrob[axis.title.y.left..zeroGrob.43]
#> 30 30 (11-11,27-27)   ylab-r-2                             zeroGrob[NULL]
#> 31 31 ( 5- 5,24-24) subtitle-2       zeroGrob[plot.subtitle..zeroGrob.45]
#> 32 32 ( 4- 4,24-24)    title-2          zeroGrob[plot.title..zeroGrob.44]
#> 33 33 (15-15,24-24)  caption-2        zeroGrob[plot.caption..zeroGrob.47]
#> 34 34 ( 3- 3,18-18)      tag-2            zeroGrob[plot.tag..zeroGrob.46]
#> 35 39 ( 9- 9,37-37)   spacer-3                             zeroGrob[NULL]
#> 36 41 (11-11,37-37)   axis-l-3         absoluteGrob[GRID.absoluteGrob.60]
#> 37 37 (13-13,37-37)   spacer-3                             zeroGrob[NULL]
#> 38 40 ( 9- 9,39-39)   axis-t-3                             zeroGrob[NULL]
#> 39 35 (11-11,39-39)    panel-3                    gTree[panel-1.gTree.58]
#> 40 43 (13-13,39-39)   axis-b-3         absoluteGrob[GRID.absoluteGrob.59]
#> 41 38 ( 9- 9,41-41)   spacer-3                             zeroGrob[NULL]
#> 42 42 (11-11,41-41)   axis-r-3                             zeroGrob[NULL]
#> 43 36 (13-13,41-41)   spacer-3                             zeroGrob[NULL]
#> 44 44 ( 8- 8,39-39)   xlab-t-3                             zeroGrob[NULL]
#> 45 45 (14-14,39-39)   xlab-b-3 zeroGrob[axis.title.x.bottom..zeroGrob.61]
#> 46 46 (11-11,36-36)   ylab-l-3   zeroGrob[axis.title.y.left..zeroGrob.62]
#> 47 47 (11-11,42-42)   ylab-r-3                             zeroGrob[NULL]
#> 48 48 ( 5- 5,39-39) subtitle-3       zeroGrob[plot.subtitle..zeroGrob.64]
#> 49 49 ( 4- 4,39-39)    title-3          zeroGrob[plot.title..zeroGrob.63]
#> 50 50 (15-15,39-39)  caption-3        zeroGrob[plot.caption..zeroGrob.66]
#> 51 51 ( 3- 3,33-33)      tag-3            zeroGrob[plot.tag..zeroGrob.65]
#> 52  0 ( 1-20, 1-47) background             rect[plot.background..rect.81]
#> 
#> $`1`$height
#> [1] 3.162278
#> 
#> $`1`$xmin
#> [1] 2
#> 
#> $`1`$xmax
#> [1] 3
#> 
#> $`1`$heightup
#> [1] 18.43909
#> 
#>